fix(page tabs): abstract page tabs layout and custom badge, fix invitation result text
