+++ client/app/bundles/course/user-invitations/components/misc/InvitationResultDialog.tsx
@@ -45,8 +45,8 @@ const translations = defineMessages({
   body: {
     id: 'course.userInvitations.components.misc.InvitationResultDialog.body',
     defaultMessage:
-      '{newInvitationsCount, plural, =0 {No new users were} one {# new user was} other {# new users were}} invited to Coursemology. ' +
-      '{newCourseUsersCount, plural, =0 {No users without Coursemology accounts were} one {# new user without Coursemology account was} other {# new users without Coursemology accounts were}} invited to this course.',
+      '{newInvitationsCount, plural, =0 {No new users were} one {# new user has been} other {# new users have been}} invited to Coursemology. ' +
+      '{newCourseUsersCount, plural, =0 {No user with Coursemology account has been} one {# new user with Coursemology account has been} other {# new users with Coursemology accounts have been}} added to this course.',
   },
   duplicateInfo: {
     id: 'course.userInvitations.components.misc.InvitationResultDialog.duplicateUsers.info',
