+++ client/app/theme/mui-style.ts
@@ -1,0 +1,12 @@
+/* eslint-disable import/prefer-default-export */
+export const tabsStyle = {
+  // to show tab indicator on firefox
+  '& .MuiTabs-indicator': {
+    bottom: 'auto',
+  },
+  '& button': { minHeight: '48px' },
+  minHeight: '50px',
+  '& .MuiTab-root:focus': {
+    outline: 0,
+  },
+};
